% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lawmakers.R
\docType{data}
\name{lawmakers}
\alias{lawmakers}
\title{Arkansas lawmakers from 2001 through the 2019 regular session.}
\format{A nested tibble with 474 rows and 3 variables:
\describe{
  \item{sponsor_full_name}{name of the lawmaker}
  \item{count}{the number of cycle/session entries for the lawmaker}
  \item{sessions}{information about the lawmaker from each session in which the lawmaker sponsored an act}
}}
\source{
\url{https://github.com/titaniumtroop/aRlegislation/}
}
\usage{
lawmakers
}
\description{
A dataset for Arkansas lawmakers. The format is a nested tibble; the
nesting variable is the lawmaker's full name. The nested data includes
detailed information about the lawmaker from each session, such as district,
seniority, committee membership, etc. For convenience, the count of the total
number of cycle/session combinations is provided.
}
\keyword{datasets}
