% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislation.R
\docType{data}
\name{legislation}
\alias{legislation}
\title{Arkansas legislation from 2001 through the 2019 regular session.}
\format{A nested tibble with 27 rows and 5 variables:
\describe{
  \item{cycle}{year of the legislative cycle}
  \item{session}{session type and number}
  \item{acts}{tibble of act number, title, subtitle, text, and partisanship}
  \item{lawmakers}{tibble of demographic information about elected representatives}
  \item{sponsorship}{tibble of act number and sponsor}
}}
\source{
\url{https://github.com/titaniumtroop/aRlegislation/}
}
\usage{
legislation
}
\description{
A dataset for Arkansas legislation. The format is a nested tibble; the
nesting variables are the beginning year of each two-year legislative cycle
and the type (and number, if applicable) of the session. There are three
tibbles within each cycle/session. The first contains information about all
acts passed during the relevant cycle/session, including the text of the act,
the title and subtitle, the act and bill number, and the partisanship of the
act (i.e., whether it was passed by a single party, a committee, or in a
bipartisan fashion). The second tibble provides demographic information about
the lawmakers who sponsored legislation during the cycle session, including
variables such as chamber, party, seniority (a negative integer, since lower
numbers are better), committee membership, district, prior public service,
veteran status, and church affiliation. The third table contains act
sponsorship information in a simple list of sponsor name and act number.
}
\examples{
  head(legislation$acts[[1]])

}
\keyword{datasets}
