% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lawmaker_session_models.R
\name{load_lawmaker_session_models}
\alias{load_lawmaker_session_models}
\title{Loads pre-generated lawmaker-session models based on LDA model topic distributions}
\usage{
load_lawmaker_session_models(
  datadir = "./data/lawmaker_session_models/",
  topics = NULL
)
}
\arguments{
\item{datadir}{A directory containing lawmaker-session models in nested dataframe format.}

\item{topics}{An integer or a vector of integers Defaults to first 10.}
}
\value{
A nested dataframe with a column for lawmaker_models for each of the topics specified.
}
\description{
Returns nested data frames containing the dot product of a sparse per-session
sponsorship matrix and the Gibbs-type LDA models generated using the
topicmodels::LDA function. These models were generated on the entire
legislative dataset from 2001-2019 on a per-session basis. Because the
generation is time-consuming, the individuals models have been saved for ease
of reloading. The models take up > 50 MB of space in the aggregate, so they
are saved individually. This function permits topics within a specified range
to be loaded.
}
