% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_topic_models.R
\name{load_topic_models}
\alias{load_topic_models}
\title{Loads pre-generated LDA topic models of aRlegislation}
\usage{
load_topic_models(datadir = NULL, topics = NULL)
}
\arguments{
\item{datadir}{A directory containing topic models in nested dataframe format. If NULL, the inst/extdata and extdata directories are searched in the installation directory. The inst/extdata directory should be explicitly declared to use all files, as only a few models are provided in the installation.}

\item{topics}{An integer, a vector of integers, or named vector in c(to = x, from = y) format. Defaults to first 10.}
}
\value{
A nested dataframe with a column named topic_models for each of the topics specified, along with scores calculated by the ldatuning package.
}
\description{
Returns Gibbs-type LDA models generated using the topicmodels::LDA function.
These topic models were generated on the entire legislative dataset from
2001-2019. Because the generation is time-consuming, the individuals models
have been saved for ease of reloading. The models take up ~15MB of space
each, so they are saved individually. This function permits topics within a
specified range to be loaded.
}
